requirements.txt
langchain
langchain-openai
langchain-community
langchain-experimental
langgraph
streamlit
pandas
matplotlib
seaborn
python-dotenv
pytest
sqlalchemy